pyproject.toml
[project]
name = "dotfiles"
version = "0.1.0"
description = "Collection of my dotfiles"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]
